fastrpc: hexagonrpcd: iobuffer: change types in decoder context
There is no real reason for the position attributes to be signed. It
does not make sense to have a negative offset in the current input
buffer, negative alignment, or negative offset in the size of the
current input buffer. A negative number of input buffers also makes no
sense. Reorganize the decoder context struct and the I/O buffer struct
so offsets and maximums are together, nothing has the unnecessary
capability of being negative, and the sizes are not wider than possible
in the format.

Also include the definition for uint32_t.

Also update the error message format in listener.c to read a 32-bit
integer.

+++ fastrpc/hexagonrpcd/iobuffer.c
@@ -34,7 +34,7 @@ static size_t consume_size(struct fastrpc_decoder_context *ctx,
 {
 	size_t segment;
 
-	segment = MIN(len, (size_t) 4 - ctx->size_off);
+	segment = MIN(len, 4 - ctx->size_off);
 	memcpy(&((char *) &ctx->size)[ctx->size_off], buf, segment);
 	ctx->size_off = (ctx->size_off + segment) % 4;
 	ctx->align = (ctx->align + segment) & 0x7;
@@ -55,7 +55,7 @@ static size_t consume_alignment(struct fastrpc_decoder_context *ctx, size_t len)
 	size_t segment = 0;
 
 	if (ctx->align) {
-		segment = MIN(len, (size_t) 8 - ctx->align);
+		segment = MIN(len, 8 - ctx->align);
 		ctx->align = (ctx->align + segment) & 0x7;
 	}
